+++ packages/ui-components/src/lib/__mocks__/MockComponent.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+	import { props } from './MockComponent';
+
+	// wheneveer the props update, we want to update the store with those
+	$: $props = $$props;
+</script>
