+++ packages/ui-components/src/lib/__mocks__/MockComponent.ts
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store';
+
+export const props = writable<unknown>();
