+++ packages/ui-components/src/lib/components/charts/TanstackLightweightChartLine.test.ts
@@ -1,16 +1,18 @@
 import { render, waitFor } from '@testing-library/svelte';
 import { test, expect, vi } from 'vitest';
 import TanstackLightweightChartLine from './TanstackLightweightChartLine.svelte';
-import { props } from '$lib/mocks/MockComponent';
-import { createResolvableQuery } from '@rainlanguage/ui-components';
+import { props } from '../../__mocks__/MockComponent';
+import { createResolvableQuery } from '../../__mocks__/queries';
 import type { UTCTimestamp } from 'lightweight-charts';
 import { get } from 'svelte/store';
 
 // Mock the LightweightChart component
-vi.mock('./LightweightChart.svelte', async () => {
-  const MockLightweightChart = (await import('$lib/mocks/MockComponent.svelte')).default;
-  return { default: MockLightweightChart };
-});
+vi.mock('@rainlanguage/ui-components', async (importOriginal) => {
+  const MockLightweightChart = (await import('../../__mocks__/MockComponent.svelte')).default;
+  return {
+    ...(await importOriginal<Record<string, unknown>>()),
+     LightweightChart: MockLightweightChart
+}});
 
 type MockData = {
   value: number;
