+++ tauri-app/src/lib/components/charts/LightweightChartHistogram.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import LightweightChart from './LightweightChart.svelte';
+  import { LightweightChart } from '@rainlanguage/ui-components';
 </script>
 
 <LightweightChart createSeries={(chart) => chart.addHistogramSeries()} {...$$props} />
